更改目录结构使其适合docker自动发布
