+++ bin/config/config.json
@@ -1,0 +1,7 @@
+{
+  "title" : "es推送工具",
+  "port" : 9102,
+  "jobList" : [
+    {"name": "example" , "filePath" : "example.json"}
+  ]
+}
