+++ config/Config.go
@@ -3,6 +3,9 @@ package config
 import (
 	"fmt"
 	"github.com/spf13/viper"
+	"main/service"
+	"main/service/errno"
+	"time"
 )
 
 type Config struct {
@@ -17,7 +20,7 @@ type JobList struct {
 }
 
 func NewConfig () Config{
-
+	Loop:
 	config := viper.New()
 	config.SetConfigName("config.json")
 	config.SetConfigType("json")
@@ -25,10 +28,19 @@ func NewConfig () Config{
 	err := config.ReadInConfig()
 
 	if err != nil {
-		panic(fmt.Errorf("Fatal error config file: %s \n", err))
+		fmt.Println(errno.SysConfigNotFind)
+		fmt.Println("正在生成默认配置文件")
+		makeDefaultConfig()
+		time.Sleep( 15 * time.Second)
+		goto Loop
 	}
+
 	configjson := Config{}
 	config.Unmarshal(&configjson)
 	return configjson
 }
 
+func makeDefaultConfig()  {
+	service.CopyFile("./config/config.json", "./webGUI/default/config/config.json")
+	service.CopyFile("./config/example.json", "./webGUI/default/config/example.json")
+}
