+++ webGUI/default/config/example.json
@@ -1,0 +1,61 @@
+{
+  "job": {
+    "setting": {
+      "speed": {
+        "channel" : 8
+      }
+    },
+    "content": {
+        "reader": {
+          "name": "mysqlreader",
+          "parameter": {
+            "connection": {
+                "jdbcUrl" : "root:pass@tcp(ip)/db_name",
+                "querySql" : "SELECT i.id,i.id AS a_id,'palnA' AS author,i.title,i.title as title_smart,i.keyword,i.source,REPLACE(GROUP_CONCAT(tn.name), ',', ' ') AS tags,p.c1,p.dl_num,p.cl_num,i.format,i.awidth,i.aheight,i.format_type,i.filesize,GREATEST(ifnull(pr.pr, 0), 1) AS pr,ifnull(d.down, 0) AS down,REPLACE(GROUP_CONCAT(tn.id), ',', ' ') AS t_id,UNIX_TIMESTAMP(p.createtime) AS create_time,ifnull(h.hide_sort_by_type, 0) AS hide_sort_by_type,i.aurl FROM t1_a p LEFT JOIN t1_ainfo i ON i.id = p.id LEFT JOIN t1_prinfo pr ON pr.id = p.id LEFT JOIN t1_tags t ON p.id = t.a_id LEFT JOIN t1_tags_name tn ON t.tname_id = tn.id LEFT JOIN t1_yesterday_download d ON p.id = d.pid LEFT JOIN t1_a_sort_by_hide h ON h.a_id = p.id WHERE p.sh = 1 and p.id >= ? and p.id <= ? GROUP BY i.id",
+                "boundarySql" : "SELECT min(id) as min,max(id) as max FROM t1_a"
+            }
+          }
+        },
+        "writer": {
+          "name": "elasticsearchwriter",
+          "parameter": {
+            "endpoint": "http://192.168.193.35:9200",
+            "accessId": "elastic",
+            "accessKey": "password",
+            "index": "t1",
+            "type": "_doc",
+            "batchSize": 10000,
+            "splitter": ",",
+            "column" : [
+              {"name": "id", "type": "id"},
+              {"name": "a_id", "type": "text"},
+              {"name": "author", "type": "text"},
+              {"name": "title", "type": "text"},
+              {"name": "title_smart", "type": "text"},
+              {"name": "keyword", "type": "text"},
+              {"name": "source", "type": "text"},
+              {"name": "tags", "type": "text"},
+              {"name": "c1", "type": "text"},
+              {"name": "dl_num", "type": "integer"},
+              {"name": "cl_num", "type": "integer"},
+              {"name": "format", "type": "text"},
+              {"name": "awidth", "type": "integer"},
+              {"name": "aheight", "type": "integer"},
+              {"name": "format_type", "type":"text"},
+              {"name": "filesize", "type": "integer"},
+              {"name": "pr", "type": "integer"},
+              {"name": "down", "type": "integer"},
+              {"name": "t_id", "type": "text"},
+              {"name": "create_time", "type": "integer"},
+              {"name": "hide_sort_by_type", "type": "text"},
+              {"name": "aurl", "type": "text"}
+            ],
+            "dsl" : "{\n  \"settings\": {\n    \"index\": {\n      \"sort.field\": \"pr\",\n      \"sort.order\": \"desc\",\n      \"store.type\": \"hybridfs\",\n      \"number_of_shards\": 1, \n      \"number_of_replicas\": 1,\n      \"similarity\" : {\n          \"default\" : {\n            \"type\" : \"BM25\",\n            \"b\": 0,\n            \"k1\": 1.2\n          }\n      }\n    }\n  },\n  \"mappings\": {\n    \"properties\": {\n      \"a_id\": {\n        \"type\": \"keyword\"\n      },\n      \"author\": {\n        \"type\": \"keyword\"\n      },\n      \"title\": {\n        \"type\": \"text\",\n        \"analyzer\": \"ik_max_word\",\n        \"search_analyzer\": \"ik_smart\"\n      },\n      \"title_smart\": {\n        \"type\": \"text\",\n        \"analyzer\": \"ik_smart\"\n      },\n      \"keyword\": {\n        \"type\": \"text\",\n        \"analyzer\": \"ik_smart\"\n      },\n      \"source\": {\n        \"type\": \"keyword\"\n      },\n      \"tags\": {\n        \"type\": \"text\",\n        \"analyzer\": \"ik_smart\"\n      },\n      \"c1\": {\n        \"type\": \"keyword\"\n      },\n      \"dl_num\": {\n        \"type\": \"integer\",\n        \"doc_values\": true\n      },\n      \"cl_num\": {\n        \"type\": \"integer\"\n      },\n      \"format\": {\n        \"type\": \"keyword\"\n      },\n      \"awidth\": {\n        \"type\": \"integer\"\n      },\n      \"aheight\": {\n        \"type\": \"integer\"\n      },\n      \"format_type\": {\n        \"type\": \"keyword\"\n      },\n      \"filesize\": {\n        \"type\": \"integer\"\n      },\n      \"pr\": {\n        \"type\": \"integer\",\n        \"doc_values\": true\n      },\n      \"down\": {\n        \"type\": \"integer\",\n        \"doc_values\": true\n      },\n      \"t_id\": {\n        \"type\": \"text\"\n      },\n      \"create_time\": {\n        \"type\": \"integer\",\n        \"doc_values\": true\n      },\n      \"hide_sort_by_type\": {\n        \"type\": \"keyword\"\n      },\n      \"aurl\": {\n        \"type\": \"keyword\"\n      }\n    }\n  }\n}"
+          }
+        }
+      }
+
+  }
+}
+
+
